[BOOK-67] refactor: 디렉토리 이동
