--- admin/src/test/kotlin/org/yapp/AdminApplicationTests.kt
@@ -1,13 +1,0 @@
-package org.yapp
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class AdminApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
