+++ admin/src/test/kotlin/org/yapp/admin/AdminApplicationTests.kt
@@ -1,13 +1,13 @@
-package org.yapp
+package org.yapp.admin
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest
-class ApisApplicationTests {
+class AdminApplicationTests {
 
-    @Test
-    fun contextLoads() {
-    }
+	@Test
+	fun contextLoads() {
+	}
 
 }
