+++ apis/src/main/kotlin/org/yapp/apis/auth/exception/AuthException.kt
@@ -1,6 +1,6 @@
 package org.yapp.apis.auth.exception
 
-import org.yapp.global.exception.CommonException
+import org.yapp.globalutils.global.exception.CommonException
 
 class AuthException(
     errorCode: AuthErrorCode,
