+++ apis/src/test/kotlin/org/yapp/apis/ApisApplicationTests.kt
@@ -1,10 +1,10 @@
-package org.yapp
+package org.yapp.apis
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
 
-@SpringBootTest(classes = [GlobalUtilsApplicationTests::class])
-class GlobalUtilsApplicationTests {
+@SpringBootTest
+class ApisApplicationTests {
 
     @Test
     fun contextLoads() {
