+++ domain/src/main/kotlin/org/yapp/domain/service/domain/UserDomainService.kt
@@ -1,11 +1,11 @@
 package org.yapp.domain.service.domain
 
-import org.yapp.annotation.DomainService
+import org.yapp.globalutils.DomainService
 import org.yapp.domain.auth.ProviderType
 import org.yapp.domain.user.User
 import org.yapp.domain.user.UserRepository
 import org.yapp.domain.user.vo.SocialUserProfile
-import org.yapp.global.util.TimeProvider
+import org.yapp.globalutils.TimeProvider
 import java.util.*
 
 @DomainService
