+++ domain/src/main/kotlin/org/yapp/domain/service/redis/TokenDomainRedisService.kt
@@ -1,6 +1,6 @@
 package org.yapp.domain.service.redis
 
-import org.yapp.annotation.DomainService
+import org.yapp.globalutils.DomainService
 import org.yapp.domain.auth.TokenRepository
 import java.util.*
