+++ domain/src/test/kotlin/org/yapp/domain/DomainApplicationTests.kt
@@ -1,0 +1,13 @@
+package org.yapp.domain
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest(classes = [DomainApplicationTests::class])
+class DomainApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+
+}
