+++ gateway/src/main/kotlin/org/yapp/gateway/jwt/exception/JwtErrorCode.kt
@@ -1,7 +1,7 @@
 package org.yapp.gateway.jwt.exception
 
 import org.springframework.http.HttpStatus
-import org.yapp.global.exception.BaseErrorCode
+import org.yapp.globalutils.global.exception.BaseErrorCode
 
 /**
  * Error codes for JWT-related errors.
