+++ gateway/src/test/kotlin/org/yapp/gateway/GatewayApplicationTests.kt
@@ -1,4 +1,4 @@
-package org.yapp
+package org.yapp.gateway
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
