+++ global-utils/src/main/kotlin/org/yapp/globalutils/exception/CommonException.kt
@@ -1,4 +1,4 @@
-package org.yapp.global.exception
+package org.yapp.globalutils.global.exception
 
 import org.springframework.web.server.ResponseStatusException
