+++ global-utils/src/main/kotlin/org/yapp/globalutils/exception/GlobalExceptionHandler.kt
@@ -1,4 +1,4 @@
-package org.yapp.global.exception
+package org.yapp.globalutils.global.exception
 
 import mu.KotlinLogging
 import org.springframework.http.ResponseEntity
