+++ infra/src/main/kotlin/org/yapp/infra/config/internal/jpa/JpaConfig.kt
@@ -6,6 +6,6 @@ import org.springframework.transaction.annotation.EnableTransactionManagement
 import org.yapp.infra.InfraBaseConfig
 
 @EnableTransactionManagement
-@EntityScan(basePackages = ["org.yapp"])
-@EnableJpaRepositories(basePackages = ["org.yapp"])
+@EntityScan(basePackages = ["org.yapp.infra"])
+@EnableJpaRepositories(basePackages = ["org.yapp.infra"])
 class JpaConfig : InfraBaseConfig
