+++ infra/src/main/kotlin/org/yapp/infra/user/entity/UserEntity.kt
@@ -6,7 +6,7 @@ import org.hibernate.annotations.JdbcTypeCode
 import org.yapp.domain.auth.ProviderType
 import org.yapp.domain.common.BaseTimeEntity
 import org.yapp.domain.user.User
-import org.yapp.global.util.UuidGenerator
+import org.yapp.globalutils.UuidGenerator
 import java.sql.Types
 import java.util.*
